moved dtos to dto project
