+++ Backend/DTO/DTO/SSOServicesDTOs.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DTO.DTO
+{
+    public class SSOServicesDTOs
+    {
+        public class LoginRequestPayload
+        {
+            [Required]
+            public string SSOUserId { get; set; }
+            [Required]
+            public string Email { get; set; }
+            [Required]
+            public long Timestamp { get; set; }
+            [Required]
+            public string Signature { get; set; }
+
+            public string PreSignatureString()
+            {
+                string acc = "";
+                acc += "ssoUserId=" + SSOUserId + ";";
+                acc += "email=" + Email + ";";
+                acc += "timestamp=" + Timestamp + ";";
+                return acc;
+            }
+        }
+    }
+}
