+++ Backend/WebApi-PointMap/WebApi-PointMap.csproj
@@ -123,7 +123,6 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="Models\UserControllerDTOs.cs" />
     <Compile Include="Models\PointControllerPOCO.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
